+++ fe-coding-task/src/housePricing/SearchHistory.tsx
@@ -1,4 +1,5 @@
 import { Box } from "@mui/material";
+
 import { HistoryType } from "./housingTypes";
 
 export const SearchHistory: React.FC<{ history: HistoryType }> = ({
@@ -7,7 +8,7 @@ export const SearchHistory: React.FC<{ history: HistoryType }> = ({
   return (
     <Box sx={{ margin: 2, maxHeight: "30vh" }}>
       {history.map((el) => (
-        <p key={el.date} style={{ wordWrap: "break-word" }}>
+        <p key={el.date} data-testid="history-entry" style={{ wordWrap: "break-word" }}>
           {JSON.stringify(el)}
         </p>
       ))}
